Added goodness_of_fit.py
Purpose: Designed to evaluate the goodness of fit for quasar datasets and the radius-luminosity relationship. The fitted parameters for all situations will be used as priors for the Markov Chain Monte Carlo (MCMC) simulations.

Note #1: The associated test function achieved a coverage rate of 99%.

Note #2: This PR added __init__.py files to every code directory (that is necessary so far).